remove opentracing and use otel instead
