+++ pkg/infra/tracing/tracing.go
@@ -3,22 +3,27 @@ package tracing
 import (
 	"context"
 	"fmt"
-	"io"
+	"net"
 	"net/http"
 	"os"
 	"strings"
+	"time"
 
-	opentracing "github.com/opentracing/opentracing-go"
-	"github.com/opentracing/opentracing-go/ext"
-	ol "github.com/opentracing/opentracing-go/log"
-	"github.com/uber/jaeger-client-go"
-	jaegercfg "github.com/uber/jaeger-client-go/config"
-	"github.com/uber/jaeger-client-go/zipkin"
+	"go.etcd.io/etcd/api/v3/version"
+	jaegerpropagator "go.opentelemetry.io/contrib/propagators/jaeger"
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/exporters/jaeger"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/sdk/resource"
+	tracesdk "go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	trace "go.opentelemetry.io/otel/trace"
 
-	"github.com/grafana/grafana/pkg/cmd/grafana-cli/logger"
+	"github.com/go-kit/log/level"
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/setting"
 )
@@ -27,6 +32,46 @@ const (
 	envJaegerAgentHost = "JAEGER_AGENT_HOST"
 	envJaegerAgentPort = "JAEGER_AGENT_PORT"
 )
+
+const (
+	jaegerExporter string = "jaeger"
+	otlpExporter   string = "otlp"
+	noopExporter   string = "noop"
+
+	jaegerPropagator string = "jaeger"
+	w3cPropagator    string = "w3c"
+)
+
+type Opentelemetry struct {
+	Enabled       string
+	Address       string
+	Propagation   string
+	customAttribs []attribute.KeyValue
+	sampler       string
+	samplerParam  float64
+
+	log log.Logger
+
+	tracerProvider tracerProvider
+	tracer         trace.Tracer
+
+	Cfg *setting.Cfg
+}
+
+type tracerProvider interface {
+	trace.TracerProvider
+
+	Shutdown(ctx context.Context) error
+}
+
+type OpentelemetrySpan struct {
+	span trace.Span
+}
+
+type EventValue struct {
+	Str string
+	Num int64
+}
 
 // Tracer defines the service used to create new spans.
 type Tracer interface {
@@ -83,7 +128,7 @@ type Span interface {
 }
 
 func ProvideService(cfg *setting.Cfg) (Tracer, error) {
-	ts, ots, err := parseSettings(cfg)
+	ots, err := ParseSettings(cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -95,43 +140,18 @@ func ProvideService(cfg *setting.Cfg) (Tracer, error) {
 
 		return nil, false
 	})
-
-	if ts.enabled {
-		return ts, ts.initJaegerGlobalTracer()
+	if err := ots.initOpentelemetryTracer(); err != nil {
+		return nil, err
 	}
-
-	return ots, ots.initOpentelemetryTracer()
+	return ots, nil
 }
 
-func parseSettings(cfg *setting.Cfg) (*Opentracing, *Opentelemetry, error) {
-	ts, err := parseSettingsOpentracing(cfg)
-	if err != nil {
-		return ts, nil, err
-	}
-	ots, err := ParseSettingsOpentelemetry(cfg)
-	return ts, ots, err
-}
-
-func parseSettingsOpentracing(cfg *setting.Cfg) (*Opentracing, error) {
-	ts := &Opentracing{
-		Cfg: cfg,
-		log: log.New("tracing"),
-	}
-	if err := ts.parseSettings(); err != nil {
-		return ts, err
-	}
-	if ts.enabled {
-		cfg.Logger.Warn("[Deprecated] the configuration setting 'tracing.jaeger' is deprecated, please use 'tracing.opentelemetry.jaeger' instead")
-	}
-	return ts, nil
-}
-
-func ParseSettingsOpentelemetry(cfg *setting.Cfg) (*Opentelemetry, error) {
+func ParseSettings(cfg *setting.Cfg) (*Opentelemetry, error) {
 	ots := &Opentelemetry{
 		Cfg: cfg,
 		log: log.New("tracing"),
 	}
-	err := ots.parseSettingsOpentelemetry()
+	err := ots.parseSettings()
 	return ots, err
 }
 
@@ -153,10 +173,6 @@ func TraceIDFromContext(c context.Context, requireSampled bool) string {
 // SpanFromContext returns the Span previously associated with ctx, or nil, if no such span could be found.
 // It is the equivalent of opentracing.SpanFromContext and trace.SpanFromContext.
 func SpanFromContext(ctx context.Context) Span {
-	// Look for both opentracing and opentelemetry spans.
-	if span := opentracing.SpanFromContext(ctx); span != nil {
-		return OpentracingSpan{span: span}
-	}
 	if span := trace.SpanFromContext(ctx); span != nil {
 		return OpentelemetrySpan{span: span}
 	}
@@ -173,219 +189,300 @@ func ContextWithSpan(ctx context.Context, span Span) context.Context {
 	return ctx
 }
 
-type Opentracing struct {
-	enabled                  bool
-	address                  string
-	customTags               map[string]string
-	samplerType              string
-	samplerParam             float64
-	samplingServerURL        string
-	log                      log.Logger
-	closer                   io.Closer
-	zipkinPropagation        bool
-	disableSharedZipkinSpans bool
-
-	Cfg *setting.Cfg
+type noopTracerProvider struct {
+	trace.TracerProvider
 }
 
-type OpentracingSpan struct {
-	span opentracing.Span
-}
-
-func (ts *Opentracing) parseSettings() error {
-	var section, err = ts.Cfg.Raw.GetSection("tracing.jaeger")
-	if err != nil {
-		return err
-	}
-
-	ts.address = section.Key("address").MustString("")
-	if ts.address == "" {
-		host := os.Getenv(envJaegerAgentHost)
-		port := os.Getenv(envJaegerAgentPort)
-		if host != "" || port != "" {
-			ts.address = fmt.Sprintf("%s:%s", host, port)
-		}
-	}
-	if ts.address != "" {
-		ts.enabled = true
-	}
-
-	ts.customTags = splitTagSettings(section.Key("always_included_tag").MustString(""))
-	ts.samplerType = section.Key("sampler_type").MustString("")
-	ts.samplerParam = section.Key("sampler_param").MustFloat64(1)
-	ts.zipkinPropagation = section.Key("zipkin_propagation").MustBool(false)
-	ts.disableSharedZipkinSpans = section.Key("disable_shared_zipkin_spans").MustBool(false)
-	ts.samplingServerURL = section.Key("sampling_server_url").MustString("")
+func (noopTracerProvider) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-func (ts *Opentracing) initJaegerCfg() (jaegercfg.Configuration, error) {
-	cfg := jaegercfg.Configuration{
-		ServiceName: "grafana",
-		Disabled:    !ts.enabled,
-		Sampler: &jaegercfg.SamplerConfig{
-			Type:              ts.samplerType,
-			Param:             ts.samplerParam,
-			SamplingServerURL: ts.samplingServerURL,
-		},
-		Reporter: &jaegercfg.ReporterConfig{
-			LogSpans:           false,
-			LocalAgentHostPort: ts.address,
-		},
-	}
-
-	_, err := cfg.FromEnv()
-	if err != nil {
-		return cfg, err
-	}
-	return cfg, nil
-}
-
-func (ts *Opentracing) initJaegerGlobalTracer() error {
-	cfg, err := ts.initJaegerCfg()
-	if err != nil {
-		return err
-	}
-
-	jLogger := &jaegerLogWrapper{logger: log.New("jaeger")}
-
-	options := []jaegercfg.Option{}
-	options = append(options, jaegercfg.Logger(jLogger))
-
-	for tag, value := range ts.customTags {
-		options = append(options, jaegercfg.Tag(tag, value))
-	}
-
-	if ts.zipkinPropagation {
-		zipkinPropagator := zipkin.NewZipkinB3HTTPHeaderPropagator()
-		options = append(options,
-			jaegercfg.Injector(opentracing.HTTPHeaders, zipkinPropagator),
-			jaegercfg.Extractor(opentracing.HTTPHeaders, zipkinPropagator),
-		)
-
-		if !ts.disableSharedZipkinSpans {
-			options = append(options, jaegercfg.ZipkinSharedRPCSpan(true))
+func (ots *Opentelemetry) parseSettings() error {
+	legacyCfg := struct {
+		Address           string
+		Tags              string
+		SamplingServerURL string
+	}{}
+	if section, err := ots.Cfg.Raw.GetSection("tracing.jaeger"); err == nil {
+		legacyCfg.Address = section.Key("address").MustString("")
+		if legacyCfg.Address == "" {
+			host, port := os.Getenv(envJaegerAgentHost), os.Getenv(envJaegerAgentPort)
+			if host != "" || port != "" {
+				legacyCfg.Address = fmt.Sprintf("%s:%s", host, port)
+			}
 		}
-	}
+		legacyCfg.Tags = section.Key("always_included_tag").MustString("")
+		legacyCfg.SamplingServerURL = section.Key("sampling_server_url").MustString("")
 
-	tracer, closer, err := cfg.NewTracer(options...)
+		ots.sampler = section.Key("sampler_type").MustString("")
+		ots.samplerParam = section.Key("sampler_param").MustFloat64(1)
+	}
+	section := ots.Cfg.Raw.Section("tracing.opentelemetry")
+	var err error
+	ots.customAttribs, err = splitCustomAttribs(section.Key("custom_attributes").MustString(legacyCfg.Tags))
 	if err != nil {
 		return err
 	}
 
-	opentracing.SetGlobalTracer(tracer)
+	section = ots.Cfg.Raw.Section("tracing.opentelemetry.jaeger")
+	ots.Enabled = noopExporter
 
-	ts.closer = closer
+	ots.Address = section.Key("address").MustString(legacyCfg.Address)
+	ots.Propagation = section.Key("propagation").MustString("")
+	if ots.Address != "" {
+		ots.Enabled = jaegerExporter
+		return nil
+	}
+
+	section = ots.Cfg.Raw.Section("tracing.opentelemetry.otlp")
+	ots.Address = section.Key("address").MustString("")
+	if ots.Address != "" {
+		ots.Enabled = otlpExporter
+	}
+	ots.Propagation = section.Key("propagation").MustString("")
 	return nil
 }
 
-func (ts *Opentracing) Run(ctx context.Context) error {
+func (ots *Opentelemetry) OTelExporterEnabled() bool {
+	return ots.Enabled == otlpExporter
+}
+
+func splitCustomAttribs(s string) ([]attribute.KeyValue, error) {
+	res := []attribute.KeyValue{}
+
+	attribs := strings.Split(s, ",")
+	for _, v := range attribs {
+		parts := strings.SplitN(v, ":", 2)
+		if len(parts) > 1 {
+			res = append(res, attribute.String(parts[0], parts[1]))
+		} else if v != "" {
+			return nil, fmt.Errorf("custom attribute malformed - must be in 'key:value' form: %q", v)
+		}
+	}
+
+	return res, nil
+}
+
+func (ots *Opentelemetry) initJaegerTracerProvider() (*tracesdk.TracerProvider, error) {
+	var ep jaeger.EndpointOption
+	// Create the Jaeger exporter: address can be either agent address (host:port) or collector URL
+	if host, port, err := net.SplitHostPort(ots.Address); err == nil {
+		ep = jaeger.WithAgentEndpoint(jaeger.WithAgentHost(host), jaeger.WithAgentPort(port))
+	} else {
+		ep = jaeger.WithCollectorEndpoint(jaeger.WithEndpoint(ots.Address))
+	}
+	exp, err := jaeger.New(ep)
+	if err != nil {
+		return nil, err
+	}
+
+	res, err := resource.New(
+		context.Background(),
+		resource.WithAttributes(
+			// TODO: why are these attributes different from ones added to the
+			// OTLP provider?
+			semconv.ServiceNameKey.String("grafana"),
+			attribute.String("environment", "production"),
+		),
+		resource.WithAttributes(ots.customAttribs...),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	sampler := tracesdk.AlwaysSample()
+	if ots.sampler == "const" || ots.sampler == "probabilistic" {
+		sampler = tracesdk.TraceIDRatioBased(ots.samplerParam)
+	}
+
+	tp := tracesdk.NewTracerProvider(
+		tracesdk.WithBatcher(exp),
+		tracesdk.WithResource(res),
+		tracesdk.WithSampler(sampler),
+	)
+
+	return tp, nil
+}
+
+func (ots *Opentelemetry) initOTLPTracerProvider() (*tracesdk.TracerProvider, error) {
+	client := otlptracegrpc.NewClient(otlptracegrpc.WithEndpoint(ots.Address), otlptracegrpc.WithInsecure())
+	exp, err := otlptrace.New(context.Background(), client)
+	if err != nil {
+		return nil, err
+	}
+
+	return initTracerProvider(exp, ots.customAttribs...)
+}
+
+func initTracerProvider(exp tracesdk.SpanExporter, customAttribs ...attribute.KeyValue) (*tracesdk.TracerProvider, error) {
+	res, err := resource.New(
+		context.Background(),
+		resource.WithAttributes(
+			semconv.ServiceNameKey.String("grafana"),
+			semconv.ServiceVersionKey.String(version.Version),
+		),
+		resource.WithAttributes(customAttribs...),
+		resource.WithProcessRuntimeDescription(),
+		resource.WithTelemetrySDK(),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	tp := tracesdk.NewTracerProvider(
+		tracesdk.WithBatcher(exp),
+		tracesdk.WithSampler(tracesdk.ParentBased(
+			tracesdk.AlwaysSample(),
+		)),
+		tracesdk.WithResource(res),
+	)
+	return tp, nil
+}
+
+func (ots *Opentelemetry) initNoopTracerProvider() (tracerProvider, error) {
+	return &noopTracerProvider{TracerProvider: trace.NewNoopTracerProvider()}, nil
+}
+
+func (ots *Opentelemetry) initOpentelemetryTracer() error {
+	var tp tracerProvider
+	var err error
+	switch ots.Enabled {
+	case jaegerExporter:
+		tp, err = ots.initJaegerTracerProvider()
+		if err != nil {
+			return err
+		}
+	case otlpExporter:
+		tp, err = ots.initOTLPTracerProvider()
+		if err != nil {
+			return err
+		}
+	default:
+		tp, err = ots.initNoopTracerProvider()
+		if err != nil {
+			return err
+		}
+	}
+
+	// Register our TracerProvider as the global so any imported
+	// instrumentation in the future will default to using it
+	// only if tracing is enabled
+	if ots.Enabled != "" {
+		otel.SetTracerProvider(tp)
+	}
+
+	propagators := []propagation.TextMapPropagator{}
+	for _, p := range strings.Split(ots.Propagation, ",") {
+		switch p {
+		case w3cPropagator:
+			propagators = append(propagators, propagation.TraceContext{}, propagation.Baggage{})
+		case jaegerPropagator:
+			propagators = append(propagators, jaegerpropagator.Jaeger{})
+		case "":
+		default:
+			return fmt.Errorf("unsupported OpenTelemetry propagator: %q", p)
+		}
+	}
+
+	switch len(propagators) {
+	case 0:
+		otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(
+			propagation.TraceContext{}, propagation.Baggage{},
+		))
+	case 1:
+		otel.SetTextMapPropagator(propagators[0])
+	default:
+		otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(propagators...))
+	}
+
+	if ots.tracerProvider == nil {
+		ots.tracerProvider = tp
+	}
+
+	ots.tracer = otel.GetTracerProvider().Tracer("component-main")
+
+	return nil
+}
+
+func (ots *Opentelemetry) Run(ctx context.Context) error {
+	otel.SetErrorHandler(otel.ErrorHandlerFunc(func(err error) {
+		err = level.Error(ots.log).Log("msg", "OpenTelemetry handler returned an error", "err", err)
+		if err != nil {
+			ots.log.Error("OpenTelemetry log returning error", err)
+		}
+	}))
 	<-ctx.Done()
 
-	if ts.closer != nil {
-		ts.log.Info("Closing tracing")
-		return ts.closer.Close()
+	ots.log.Info("Closing tracing")
+	if ots.tracerProvider == nil {
+		return nil
+	}
+	ctxShutdown, cancel := context.WithTimeout(ctx, time.Second*5)
+	defer cancel()
+
+	if err := ots.tracerProvider.Shutdown(ctxShutdown); err != nil {
+		return err
 	}
 
 	return nil
 }
 
-func (ts *Opentracing) Start(ctx context.Context, spanName string, opts ...trace.SpanStartOption) (context.Context, Span) {
-	span, ctx := opentracing.StartSpanFromContext(ctx, spanName)
-	opentracingSpan := OpentracingSpan{span: span}
-	if sctx, ok := span.Context().(jaeger.SpanContext); ok {
-		ctx = context.WithValue(ctx, traceKey{}, traceValue{sctx.TraceID().String(), sctx.IsSampled()})
+func (ots *Opentelemetry) Start(ctx context.Context, spanName string, opts ...trace.SpanStartOption) (context.Context, Span) {
+	ctx, span := ots.tracer.Start(ctx, spanName, opts...)
+	opentelemetrySpan := OpentelemetrySpan{
+		span: span,
 	}
-	return ctx, opentracingSpan
-}
 
-func (ts *Opentracing) Inject(ctx context.Context, header http.Header, span Span) {
-	opentracingSpan, ok := span.(OpentracingSpan)
-	if !ok {
-		logger.Error("Failed to cast opentracing span")
+	if traceID := span.SpanContext().TraceID(); traceID.IsValid() {
+		ctx = context.WithValue(ctx, traceKey{}, traceValue{traceID.String(), span.SpanContext().IsSampled()})
 	}
-	err := opentracing.GlobalTracer().Inject(
-		opentracingSpan.span.Context(),
-		opentracing.HTTPHeaders,
-		opentracing.HTTPHeadersCarrier(header))
 
-	if err != nil {
-		logger.Error("Failed to inject span context instance", "err", err)
-	}
+	return ctx, opentelemetrySpan
 }
 
-func (s OpentracingSpan) End() {
-	s.span.Finish()
+func (ots *Opentelemetry) Inject(ctx context.Context, header http.Header, _ Span) {
+	otel.GetTextMapPropagator().Inject(ctx, propagation.HeaderCarrier(header))
 }
 
-func (s OpentracingSpan) SetAttributes(key string, value interface{}, kv attribute.KeyValue) {
-	s.span.SetTag(key, value)
+func (s OpentelemetrySpan) End() {
+	s.span.End()
 }
 
-func (s OpentracingSpan) SetName(name string) {
-	s.span.SetOperationName(name)
+func (s OpentelemetrySpan) SetAttributes(key string, value interface{}, kv attribute.KeyValue) {
+	s.span.SetAttributes(kv)
 }
 
-func (s OpentracingSpan) SetStatus(code codes.Code, description string) {
-	if code == codes.Error {
-		ext.Error.Set(s.span, true)
-	}
+func (s OpentelemetrySpan) SetName(name string) {
+	s.span.SetName(name)
 }
 
-func (s OpentracingSpan) RecordError(err error, options ...trace.EventOption) {
-	ext.Error.Set(s.span, true)
+func (s OpentelemetrySpan) SetStatus(code codes.Code, description string) {
+	s.span.SetStatus(code, description)
 }
 
-func (s OpentracingSpan) AddEvents(keys []string, values []EventValue) {
-	fields := []ol.Field{}
+func (s OpentelemetrySpan) RecordError(err error, options ...trace.EventOption) {
+	s.span.RecordError(err, options...)
+}
+
+func (s OpentelemetrySpan) AddEvents(keys []string, values []EventValue) {
 	for i, v := range values {
 		if v.Str != "" {
-			field := ol.String(keys[i], v.Str)
-			fields = append(fields, field)
+			s.span.AddEvent(keys[i], trace.WithAttributes(attribute.Key(keys[i]).String(v.Str)))
 		}
 		if v.Num != 0 {
-			field := ol.Int64(keys[i], v.Num)
-			fields = append(fields, field)
+			s.span.AddEvent(keys[i], trace.WithAttributes(attribute.Key(keys[i]).Int64(v.Num)))
 		}
 	}
-	s.span.LogFields(fields...)
 }
 
-func (s OpentracingSpan) contextWithSpan(ctx context.Context) context.Context {
+func (s OpentelemetrySpan) contextWithSpan(ctx context.Context) context.Context {
 	if s.span != nil {
-		ctx = opentracing.ContextWithSpan(ctx, s.span)
+		ctx = trace.ContextWithSpan(ctx, s.span)
 		// Grafana also manages its own separate traceID in the context in addition to what opentracing handles.
 		// It's derived from the span. Ensure that we propagate this too.
-		if sctx, ok := s.span.Context().(jaeger.SpanContext); ok {
-			ctx = context.WithValue(ctx, traceKey{}, traceValue{sctx.TraceID().String(), sctx.IsSampled()})
+		if traceID := s.span.SpanContext().TraceID(); traceID.IsValid() {
+			ctx = context.WithValue(ctx, traceKey{}, traceValue{traceID.String(), s.span.SpanContext().IsSampled()})
 		}
 	}
 	return ctx
-}
-
-func splitTagSettings(input string) map[string]string {
-	res := map[string]string{}
-
-	tags := strings.Split(input, ",")
-	for _, v := range tags {
-		kv := strings.Split(v, ":")
-		if len(kv) > 1 {
-			res[kv[0]] = kv[1]
-		}
-	}
-
-	return res
-}
-
-type jaegerLogWrapper struct {
-	logger log.Logger
-}
-
-func (jlw *jaegerLogWrapper) Error(msg string) {
-	jlw.logger.Error(msg)
-}
-
-func (jlw *jaegerLogWrapper) Infof(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	jlw.logger.Info(msg)
 }
